styles li's and renames breakpoints to utilities, because now im gonna store any css helpers and stuff into there.
