+++ js/utilities.js
@@ -1,5 +1,12 @@
 import { css } from 'styled-components'; // this has to be in brackets, apparently it has two css vars
 
+// color helpers
+export const colors = {
+  lightBlue: 'rgba(146, 220, 229, 0.75)',
+  blueShadow: 'rgba(146, 220, 229, 1)',
+  puce: 'rgba(170, 91, 97, 0.75)'
+};
+
 const sizes = {
   desktop: 992,
   tablet: 768,
@@ -13,5 +20,4 @@ const media = Object.keys(sizes).reduce((acc, label) => {
 }, {});
 
 export default media;
-
